Revert 1fa2971 (breaking group change in `order`)
1fa2971 changed the way groups work when there is only one, leading
to single nested groups being treated as though they were unnested.
Prior to this, if you wanted to group e.g. builtin and external
imports together at the top and everything else together as a second
group, a single nested group was the way to do it

[It appears that this change was unintentional][1], and was made to
try to fix what seems to be a misunderstanding around nested groups
([#2687]). [The docs][2] continue to suggest that nested groups
should be "mingled together" and makes no reference to a single nested
group with no other groups being an invalid option

This therefore reverts the change to how groups work when there is
only one. No documentation change should be necessary given this is
already the described behaviour

[1]: https://github.com/import-js/eslint-plugin-import/issues/2687#issuecomment-1671038558
[2]: https://github.com/import-js/eslint-plugin-import/blob/main/docs/rules/order.md#groups-array