fix: pnp issue when used in a monorepo
Always use the file being linted as the "location" of the require when
loading a resolver for the file. This prevents an issue when the sourceFile
 is in a dependency of the package being linted.
Add a fallback when loading a parser, the fallback uses the file being
linted as the location of the require.

Using the file being linted isn't perfect because a user might configure
the linter to lint files in other packages. However the approach is better
than what was before.